+++ src/stores/disk.rs
@@ -11,8 +11,10 @@ use std::{fmt::Display, path::PathBuf, time::SystemTime};
 pub struct DiskCacheBuilder<K, V> {
     seconds: Option<u64>,
     refresh: bool,
+    sync_to_disk_on_cache_change: bool,
     disk_dir: Option<PathBuf>,
     cache_name: String,
+    connection_config: Option<sled::Config>,
     _phantom: PhantomData<(K, V)>,
 }
 
@@ -42,8 +44,10 @@ where
         Self {
             seconds: None,
             refresh: false,
+            sync_to_disk_on_cache_change: false,
             disk_dir: None,
             cache_name: cache_name.as_ref().to_string(),
+            connection_config: None,
             _phantom: Default::default(),
         }
     }
@@ -63,6 +67,41 @@ where
     /// Set the disk path for where the data will be stored
     pub fn set_disk_directory<P: AsRef<Path>>(mut self, dir: P) -> Self {
         self.disk_dir = Some(dir.as_ref().into());
+        self
+    }
+
+    /// Specify whether the cache should sync to disk on each cache change.
+    /// [sled] flushes every [sled::Config::flush_every_ms] which has a default value.
+    /// In some use cases, the default value may not be quick enough,
+    /// or a user may want to reduce the flush rate / turn off auto-flushing to reduce IO (and only flush on cache changes).
+    /// (see [DiskCacheBuilder::set_connection_config] for more control over the sled connection)
+    pub fn set_sync_to_disk_on_cache_change(mut self, sync_to_disk_on_cache_change: bool) -> Self {
+        self.sync_to_disk_on_cache_change = sync_to_disk_on_cache_change;
+        self
+    }
+
+    /// Specify the [sled::Config] to use for the connection to the disk cache.
+    ///
+    /// ### Note
+    /// Don't use [sled::Config::path] as any value set here will be overwritten by either
+    /// the path specified in [DiskCacheBuilder::set_disk_directory], or the default value calculated by [DiskCacheBuilder].
+    ///
+    /// ### Example Use Case
+    /// By default [sled] automatically syncs to disk at a frequency specified in [sled::Config::flush_every_ms].
+    /// A user may want to reduce IO by setting a lower flush frequency, or by setting [sled::Config::flush_every_ms] to [None].
+    /// Also see [DiskCacheBuilder::set_sync_to_disk_on_cache_change] which allows for syncing to disk on each cache change.
+    /// ```rust
+    /// use cached::stores::{DiskCacheBuilder, DiskCache};
+    ///
+    /// let config = sled::Config::new().flush_every_ms(None);
+    /// let cache: DiskCache<String, String> = DiskCacheBuilder::new("my-cache")
+    ///     .set_connection_config(config)
+    ///     .set_sync_to_disk_on_cache_change(true)
+    ///     .build()
+    ///     .unwrap();
+    /// ```
+    pub fn set_connection_config(mut self, config: sled::Config) -> Self {
+        self.connection_config = Some(config);
         self
     }
 
@@ -87,11 +126,15 @@ where
     pub fn build(self) -> Result<DiskCache<K, V>, DiskCacheBuildError> {
         let disk_dir = self.disk_dir.unwrap_or_else(|| Self::default_disk_dir());
         let disk_path = disk_dir.join(format!("{}_v{}", self.cache_name, DISK_FILE_VERSION));
-        let connection = sled::open(disk_path.clone())?;
+        let connection = match self.connection_config {
+            Some(config) => config.path(disk_path.clone()).open()?,
+            None => sled::open(disk_path.clone())?,
+        };
 
         Ok(DiskCache {
             seconds: self.seconds,
             refresh: self.refresh,
+            sync_to_disk_on_cache_change: self.sync_to_disk_on_cache_change,
             version: DISK_FILE_VERSION,
             disk_path,
             connection,
@@ -104,6 +147,7 @@ where
 pub struct DiskCache<K, V> {
     pub(super) seconds: Option<u64>,
     pub(super) refresh: bool,
+    sync_to_disk_on_cache_change: bool,
     #[allow(unused)]
     version: u64,
     #[allow(unused)]
@@ -123,7 +167,7 @@ where
         DiskCacheBuilder::new(cache_name)
     }
 
-    pub fn remove_expired_entries(&self) {
+    pub fn remove_expired_entries(&self) -> Result<(), DiskCacheError> {
         let now = SystemTime::now();
 
         for (key, value) in self.connection.iter().flatten() {
@@ -134,11 +178,27 @@ where
                         .unwrap_or(Duration::from_secs(0))
                         >= Duration::from_secs(lifetime_seconds)
                     {
-                        let _ = self.connection.remove(key);
+                        self.connection.remove(key)?;
                     }
                 }
             }
         }
+
+        if self.sync_to_disk_on_cache_change {
+            self.connection.flush()?;
+        }
+        Ok(())
+    }
+
+    /// Provide access to the underlying [sled::Db] connection
+    /// This is useful for i.e. manually flushing the cache to disk.
+    pub fn connection(&self) -> &Db {
+        &self.connection
+    }
+
+    /// Provide mutable access to the underlying [sled::Db] connection
+    pub fn connection_mut(&mut self) -> &mut Db {
+        &mut self.connection
     }
 }
 
@@ -184,6 +244,7 @@ where
         let key = key.to_string();
         let seconds = self.seconds;
         let refresh = self.refresh;
+        let mut cache_updated = false;
         let update = |old: Option<&[u8]>| -> Option<Vec<u8>> {
             let old = old?;
             if seconds.is_none() {
@@ -204,6 +265,7 @@ where
             {
                 if refresh {
                     cached.refresh_created_at();
+                    cache_updated = true;
                 }
                 let cache_val =
                     rmp_serde::to_vec(&cached).expect("error serializing cached disk value");
@@ -213,19 +275,25 @@ where
             }
         };
 
-        if let Some(data) = self.connection.update_and_fetch(key, update)? {
+        let result = if let Some(data) = self.connection.update_and_fetch(key, update)? {
             let cached = rmp_serde::from_slice::<CachedDiskValue<V>>(&data)?;
             Ok(Some(cached.value))
         } else {
             Ok(None)
+        };
+
+        if cache_updated && self.sync_to_disk_on_cache_change {
+            self.connection.flush()?;
         }
+
+        result
     }
 
     fn cache_set(&self, key: K, value: V) -> Result<Option<V>, DiskCacheError> {
         let key = key.to_string();
         let value = rmp_serde::to_vec(&CachedDiskValue::new(value))?;
 
-        if let Some(data) = self.connection.insert(key, value)? {
+        let result = if let Some(data) = self.connection.insert(key, value)? {
             let cached = rmp_serde::from_slice::<CachedDiskValue<V>>(&data)?;
 
             if let Some(lifetime_seconds) = self.seconds {
@@ -243,12 +311,18 @@ where
             }
         } else {
             Ok(None)
+        };
+
+        if self.sync_to_disk_on_cache_change {
+            self.connection.flush()?;
         }
+
+        result
     }
 
     fn cache_remove(&self, key: &K) -> Result<Option<V>, DiskCacheError> {
         let key = key.to_string();
-        if let Some(data) = self.connection.remove(key)? {
+        let result = if let Some(data) = self.connection.remove(key)? {
             let cached = rmp_serde::from_slice::<CachedDiskValue<V>>(&data)?;
 
             if let Some(lifetime_seconds) = self.seconds {
@@ -266,7 +340,13 @@ where
             }
         } else {
             Ok(None)
+        };
+
+        if self.sync_to_disk_on_cache_change {
+            self.connection.flush()?;
         }
+
+        result
     }
 
     fn cache_lifespan(&self) -> Option<u64> {
@@ -284,6 +364,10 @@ where
         self.refresh = refresh;
         old
     }
+
+    fn cache_unset_lifespan(&mut self) -> Option<u64> {
+        self.seconds.take()
+    }
 }
 
 #[cfg(test)]
@@ -300,10 +384,19 @@ mod test_DiskCache {
 
     use super::*;
 
+    /// If passing `no_exist` to the macro:
+    /// This gives you a TempDir where the directory does not exist
+    /// so you can copy / move things to the returned TmpDir.path()
+    /// and those files will be removed when the TempDir is dropped
     macro_rules! temp_dir {
         () => {
             TempDir::new().expect("Error creating temp dir")
         };
+        (no_exist) => {{
+            let tmp_dir = TempDir::new().expect("Error creating temp dir");
+            std::fs::remove_dir_all(tmp_dir.path()).expect("error emptying the tmp dir");
+            tmp_dir
+        }};
     }
 
     fn now_millis() -> u128 {
@@ -320,7 +413,7 @@ mod test_DiskCache {
     const LIFE_SPAN_2_SECS: u64 = 2;
     const LIFE_SPAN_1_SEC: u64 = 1;
 
-    #[test]
+    #[googletest::test]
     fn cache_get_after_cache_remove_returns_none() {
         let tmp_dir = temp_dir!();
         let cache: DiskCache<u32, u32> = DiskCache::new("test-cache")
@@ -365,7 +458,7 @@ mod test_DiskCache {
         drop(cache);
     }
 
-    #[test]
+    #[googletest::test]
     fn values_expire_when_lifespan_elapses_returning_none() {
         let tmp_dir = temp_dir!();
         let cache: DiskCache<u32, u32> = DiskCache::new("test-cache")
@@ -401,7 +494,7 @@ mod test_DiskCache {
         );
     }
 
-    #[test]
+    #[googletest::test]
     fn set_lifespan_to_a_different_lifespan_is_respected() {
         // COPY PASTE of [values_expire_when_lifespan_elapses_returning_none]
         let tmp_dir = temp_dir!();
@@ -488,7 +581,7 @@ mod test_DiskCache {
         );
     }
 
-    #[test]
+    #[googletest::test]
     fn refreshing_on_cache_get_delays_cache_expiry() {
         // NOTE: Here we're relying on the fact that setting then sleeping for 2 secs and getting takes longer than 2 secs.
         const LIFE_SPAN: u64 = 2;
@@ -530,7 +623,7 @@ mod test_DiskCache {
         drop(cache);
     }
 
-    #[test]
+    #[googletest::test]
     // TODO: Consider removing this test, as it's not really testing anything.
     // If we want to check that setting a different disk directory to the default doesn't change anything,
     // we should design the tests to run all the same tests but paramaterized with different conditions.
@@ -560,5 +653,177 @@ mod test_DiskCache {
 
         // remove the cache dir to clean up the test as we're not using a temp dir
         std::fs::remove_dir_all(cache.disk_path).expect("error in clean up removeing the cache dir")
+    }
+
+    mod set_sync_to_disk_on_cache_change {
+
+        mod when_no_auto_flushing {
+            use super::super::*;
+
+            fn check_on_recovered_cache(
+                set_sync_to_disk_on_cache_change: bool,
+                run_on_original_cache: fn(&DiskCache<u32, u32>) -> (),
+                run_on_recovered_cache: fn(&DiskCache<u32, u32>) -> (),
+            ) {
+                let original_cache_tmp_dir = temp_dir!();
+                let copied_cache_tmp_dir = temp_dir!(no_exist);
+                const CACHE_NAME: &str = "test-cache";
+
+                let cache: DiskCache<u32, u32> = DiskCache::new(CACHE_NAME)
+                    .set_disk_directory(original_cache_tmp_dir.path())
+                    .set_sync_to_disk_on_cache_change(set_sync_to_disk_on_cache_change) // WHAT'S BEING TESTED
+                    // NOTE: disabling automatic flushing, so that we only test the flushing of cache_set
+                    .set_connection_config(sled::Config::new().flush_every_ms(None))
+                    .build()
+                    .unwrap();
+
+                // flush the cache to disk before any cache setting, so that when we create the recovered cache
+                // it has something to recover from, even if set_cache doesn't write to disk as we'd like.
+                cache
+                    .connection
+                    .flush()
+                    .expect("error flushing cache before any cache setting");
+
+                run_on_original_cache(&cache);
+
+                // freeze the current state of the cache files by copying them to a new location
+                // we do this before dropping the cache, as dropping the cache seems to flush to the disk
+                let recovered_cache = clone_cache_to_new_location_no_flushing(
+                    CACHE_NAME,
+                    &cache,
+                    copied_cache_tmp_dir.path(),
+                );
+
+                assert_that!(recovered_cache.connection.was_recovered(), eq(true));
+
+                run_on_recovered_cache(&recovered_cache);
+            }
+
+            mod changes_persist_after_recovery_when_set_to_true {
+                use super::*;
+
+                #[googletest::test]
+                fn for_cache_set() {
+                    check_on_recovered_cache(
+                        false,
+                        |cache| {
+                            // write to the cache, we expect this to persist if the connection is flushed on cache_set
+                            cache
+                                .cache_set(TEST_KEY, TEST_VAL)
+                                .expect("error setting cache in assemble stage");
+                        },
+                        |recovered_cache| {
+                            assert_that!(
+                                    recovered_cache.cache_get(&TEST_KEY),
+                                    ok(none()),
+                                    "set_sync_to_disk_on_cache_change is false, and there is no auto-flushing, so the cache should not have persisted"
+                                );
+                        },
+                    )
+                }
+
+                #[googletest::test]
+                fn for_cache_remove() {
+                    check_on_recovered_cache(
+                        false,
+                        |cache| {
+                            // write to the cache, we expect this to persist if the connection is flushed on cache_set
+                            cache
+                                .cache_set(TEST_KEY, TEST_VAL)
+                                .expect("error setting cache in assemble stage");
+
+                            // manually flush the cache so that we only test cache_remove
+                            cache.connection.flush().expect("error flushing cache");
+
+                            cache
+                                .cache_remove(&TEST_KEY)
+                                .expect("error removing cache in assemble stage");
+                        },
+                        |recovered_cache| {
+                            assert_that!(
+                                    recovered_cache.cache_get(&TEST_KEY),
+                                    ok(some(eq(TEST_VAL))),
+                                    "set_sync_to_disk_on_cache_change is false, and there is no auto-flushing, so the cache_remove should not have persisted"
+                                );
+                        },
+                    )
+                }
+
+                #[ignore = "Not implemented"]
+                #[googletest::test]
+                fn for_cache_get_when_refreshing() {
+                    todo!("Test not implemented.")
+                }
+            }
+
+            /// This is the anti-test
+            mod changes_do_not_persist_after_recovery_when_set_to_false {
+                use super::*;
+
+                #[googletest::test]
+                fn for_cache_set() {
+                    check_on_recovered_cache(
+                        true,
+                        |cache| {
+                            // write to the cache, we expect this to persist if the connection is flushed on cache_set
+                            cache
+                                .cache_set(TEST_KEY, TEST_VAL)
+                                .expect("error setting cache in assemble stage");
+                        },
+                        |recovered_cache| {
+                            assert_that!(
+                                recovered_cache.cache_get(&TEST_KEY),
+                                ok(some(eq(TEST_VAL))),
+                                "Getting a set key should return the value"
+                            );
+                        },
+                    )
+                }
+
+                #[googletest::test]
+                fn for_cache_remove() {
+                    check_on_recovered_cache(
+                        true,
+                        |cache| {
+                            // write to the cache, we expect this to persist if the connection is flushed on cache_set
+                            cache
+                                .cache_set(TEST_KEY, TEST_VAL)
+                                .expect("error setting cache in assemble stage");
+
+                            cache
+                                .cache_remove(&TEST_KEY)
+                                .expect("error removing cache in assemble stage");
+                        },
+                        |recovered_cache| {
+                            assert_that!(
+                                recovered_cache.cache_get(&TEST_KEY),
+                                ok(none()),
+                                "Getting a removed key should return None"
+                            );
+                        },
+                    )
+                }
+
+                #[ignore = "Not implemented"]
+                #[googletest::test]
+                fn for_cache_get_when_refreshing() {
+                    todo!("Test not implemented.")
+                }
+            }
+
+            fn clone_cache_to_new_location_no_flushing(
+                cache_name: &str,
+                cache: &DiskCache<u32, u32>,
+                new_location: &Path,
+            ) -> DiskCache<u32, u32> {
+                copy_dir::copy_dir(cache.disk_path.parent().unwrap(), new_location)
+                    .expect("error copying cache files to new location");
+
+                DiskCache::new(cache_name)
+                    .set_disk_directory(new_location)
+                    .build()
+                    .expect("error building cache from copied files")
+            }
+        }
     }
 }